fix: remove slow network test
There's no reliable way to test this in a transport agnostic way.

The idea is if async work is done after `.dial` is invoked but
before `upgrader.upgradeInbound` is called, and the abort signal
aborts, then `.dial` should respect the abort and throw, but there's
no cross-transport way to inject the latency that would cause the
test to fail so remove the test.