fix: ignore IPv6 link-local addresses (#2865)
Network interfaces generate their own link-local addresses when
bootstrapping to the network.  With IPv4 these are discarded when
a DHCP server assigns an address but IPv6 interfaces keep them
around.

These aren't useful addresses after a routable address has been
assigned so do not return them as addresses we are listening on.