fix: upgrade observed address to ip mapping (#2941)
When an observed address with a publicly routable IPv4 host is confirmed by autonat, and there's only one listener it could be associated with, upgrade the observed address to an IP mapping, this allows the DNS mapping to add SNI tuples to corresponding transport addresses with a TLS tuple.

This will, for example, detect when external infrastructure such as a router has been configured manually (e.g. not by UPnP) to forward traffic to the current node.

Fixes #2929