feat: expose maxDataLength for outgoing messages in PeerStreams (#2954)
This PR introduces a new feature that allows users to configure the `maxDataLength` option used by `it-length-prefixed` through the `PeerStreams` options. Previously, the library defaulted to a maximum data length of 4MB, which could lead to issues when processing larger messages. With this change, both the encoder and decoder in `it-length-prefixed` can be configured to use a custom `maxDataLength` thus accommodating larger messages.

A new test has been added to validate the behaviour, ensuring that a 6MB message is correctly transmitted and received without triggering an error, and that the content of the sent and received messages are identical.

---------

Co-authored-by: Russell Dempsey <1173416+SgtPooki@users.noreply.github.com>
Co-authored-by: Alex Potsides <alex@achingbrain.net>