fix: add DER decoder and allow passing protobuf digest
Adds enough of a DER decoder to parse public keys from PKIX bytes.

Also allows passing the digest of a public key in to `publicKeyFromProtobuf`
so we can avoid having to recalculate it.

This becomes a big bottleneck when loading a routing table's worth
of peers from the peer store where many of them are RSA peers.