fix: only check for non-wildcard ports (#3050)
If we are passed a wildcard port, start a UDP mux listener on a
ephemeral port, otherwise check the existing listeners before trying
to start a new one.

To be reused an existing listener must be owned by the current peer
and be of the other IP family (e.g. you cannot listen on
`[ip4-host]:1234` twice but you can listen on `[ip4-host:1234]` and
`[ip6-host]:1234`.

Fixes #3049