fix: do not dial bootstrap nodes
Updates the bootstrap module to only discover the bootstrap nodes
but not dial them.

Invidiual protocols that require bootstrapping (e.g. KAD-DHT) can
then react to the disovery and choose to dial the peers.

This cuts down on unecessary dials for lightweight nodes and brings
the bootstrap module in line with other peer discovery mechanisms.