fix: consolidate iterable closing
We have `MultiaddrConnection`s and `Stream`s, both of which have
`.source` and `.sink` properties, can be aborted or closed gracefully
by iterables ending or explicitly.

All `Stream` implementations extend `AbstractStream` which has let
us concentrate the logic that controls these (many) closing criteria
but the same is not true of `MultiaddrConnection`s.

Where we have these we transform them to async iterables which may not
suit the underlying transport.

Here we consolidate the common parts of `MultiaddrConnection`s into an
`AbstractMultiaddrConnection`, and consolidate the async iterable/closing
parts of both `MultiaddrConnection`s and `Stream`s into a
`HalfCloseableDuplex` abstract superclass.

This greatly simplifies the transport implementations and allows all
transports to benefit from tests run on the super class.