fix: ignore error event in test
Ignore the error event as it causes `raceEvent` to throw. In future
we should remove the error event from the queue as all we do with it
is log the error, and if we want that, the `failure` event gives
more context around the error so can make the logs more informative.